add cmd for Chowkidar
