--- configs/README.md
@@ -1,5 +1,0 @@
-# `/configs`
-
-Configuration file templates or default configs.
-
-Put your `confd` or `consule-template` template files here.
