--- deployments/README.md
@@ -1,3 +1,0 @@
-# `/deployments`
-
-IaaS, PaaS, system and container orchestration deployment configurations and templates (docker-compose, kubernetes/helm, mesos, terraform, bosh).
