--- deployments/kubernetes/manifests/README.md
@@ -1,2 +1,0 @@
-# kubernetes manifests
-
