+++ glide.yaml
@@ -12,3 +12,6 @@ import:
 - package: github.com/asiyani/slack
 # - package: github.com/nlopes/slack
 - package: github.com/mitchellh/mapstructure
+- package: github.com/spf13/cobra
+- package: github.com/spf13/pflag
+  version: 583c0c0531f06d5278b7d917446061adc344b5cd
