--- internal/README.md
@@ -1,5 +1,0 @@
-# `/internal`
-
-Private application and library code. This is the code you don't want others importing in their applications or libraries.
-
-Put your actual application code in the `/internal/app` directory (e.g., `/internal/app/myapp`) and the code shared by those apps in the `/internal/pkg` directory (e.g., `/internal/pkg/myprivlib`).
