+++ internal/pkg/actions/action.go
@@ -3,8 +3,8 @@ package actions
 import (
 	"log"
 
-	"github.com/stakater/Chowkidar/pkg/actions/slack"
-	"github.com/stakater/Chowkidar/pkg/config"
+	"github.com/stakater/Chowkidar/internal/pkg/actions/slack"
+	"github.com/stakater/Chowkidar/internal/pkg/config"
 	"k8s.io/api/core/v1"
 )
