+++ internal/pkg/actions/populate_test.go
@@ -4,7 +4,7 @@ import (
 	"reflect"
 	"testing"
 
-	"github.com/stakater/Chowkidar/pkg/config"
+	"github.com/stakater/Chowkidar/internal/pkg/config"
 )
 
 func TestPopulateFromConfig(t *testing.T) {
