--- pkg/README.md
@@ -1,5 +1,0 @@
-# `/pkg`
-
-Library code that's safe to use by external applications (e.g., `/pkg/mypubliclib`).
-
-Other projects will import these libraries expecting them to work, so think twice before you put something here :-)
